// src/common/enums/errorType.enum.ts
export enum ValidationErrorTypeEnum {
  DEFAULT = 100,
  IS_UUID = 101,
  IS_STRING = 102,
  IS_ARRAY = 103,
  IS_EMAIL = 104,
  IS_HEX = 105,
  IS_BOOLEAN = 106,
  IS_OBJECT = 107,
  IS_NUMBER = 108,
  IS_ENUM = 109,
  IS_NOT_EMPTY = 110,
  IS_DEFINED = 111,
  IS_NOT_EMPTY_OBJECT = 112,
  IS_LENGTH = 113,
  MIN = 114,
  MAX = 115,
  ARRAY_MIN_SIZE = 116,
  ARRAY_MAX_SIZE = 117,
  IS_VALID_REGEX = 118,
  IS_TYPE = 119,
  IS_IN = 120,
  IS_UNIQUE = 121,
  IS_STRING_OR_NULL = 122,
  DUPLICATE_VALUE = 123,
  IS_VALID_ROW_ELEMENT = 124,
  WHITE_LIST_VALIDATION = 125, // property should not exist
  IS_NOT_EMPTY_ARRAY = 126,
  IS_DATE = 127,
  IS_PHONE = 128,
  IS_URL = 129,
  NOT_FOUND = 130,
  IS_MONGO_ID = 131,
  IS_NULL = 132,
  IS_VALID_ACCESS_CONDITION_OPERATOR = 133,
  IS_VALID_ACCESS_CONDITION_VALUE = 134,
  IS_AFTER_NOW = 135,
  IS_VALID_VARIABLE_ARRAYS_LENGTH = 136,
  IS_UNIQUE_ARRAY_OF_STRING = 137,
  IS_NOT_BLANK = 138,
  IS_EXISTS_TITLE_OR_CONTENT_IN_OBJECT = 139,
  IS_POSITIVE = 140,
  IS_INT = 141,
  SHOULD_OR_NOT_EXISTS_BY_PROPERTY = 142,
}

export enum CustomErrorTypeEnum {
  DATA_NOT_FOUND = 1001,
  INCORRECT_CREDENTIALS = 1002,
  NO_ACCESS_OR_DATA_NOT_FOUND = 1003,
  ENTRY_ALREADY_EXISTS = 1004,
  ROW_POSITION_MORE_THAN_ALLOWED = 1005,
  SLIDER_NOT_ALLOWED_FOR_THIS_ROW = 1006, // row max col. should not be bigger than 1
  POSITION_ALREADY_EXISTS = 1007,
  LIMIT_FOR_ELEMENTS_IN_ROW = 1008,
  TYPE_OF_ELEMENT_NOT_ALLOWED = 1009,
  CAN_NOT_DELETE_LAST_SLIDE = 1010,
  INVALID_USER_CODE = 1011,
  USER_ALREADY_VERIFIED = 1012,
  TOO_MANY_REQUESTS = 1013,
  EXPERIMENT_ALREADY_HAS_THIS_STATUS = 1014,
  SLIDE_SHOULD_NOT_BE_PART_OF_CYCLE = 1015,
  ATTRIBUTES_LENGTH_CONFLICT = 1016, // attributes length and attributesName length did not match
  CAN_NOT_SET_TRAINING_FOR_CHILD_SLIDE = 1017,
  ARRAY_OF_ROW_DATA_NOT_UNIQUE = 1018,
  ATTRIBUTE_NOT_FOUND = 1019,
  VARIABLE_NOT_FOUND = 1020,
  ACCOUNT_NOT_VERIFIED = 1021,
  ID_NOT_FOUND = 1022,
  SLIDE_ID_NOT_FOUND = 1023,
  SLIDE_POSITION_NOT_FOUND = 1024,

  /* Experiment: start - GENERAL */
  EXPERIMENT_START_INCORRECT_USER_DATA = 1025,
  ACCESS_DENIED_FOR_THIS_PLATFORM = 1026,
  EXPERIMENT_START_CAN_NOT_START_UNPUBLISHED = 1027,

  /* Experiment: start - AGE */
  EXPERIMENT_START_AGE_EQUALS = 1028,
  EXPERIMENT_START_AGE_NOT_EQUALS = 1029,
  EXPERIMENT_START_AGE_MORE = 1030,
  EXPERIMENT_START_AGE_LESS = 1031,
  EXPERIMENT_START_AGE_MORE_OR_EQUALS = 1032,
  EXPERIMENT_START_AGE_LESS_OR_EQUALS = 1033,

  /* Experiment: start - SEX */
  EXPERIMENT_START_SEX_EQUALS = 1034,
  EXPERIMENT_START_SEX_NOT_EQUALS = 1035,

  /* Experiment: start - Native languages */
  EXPERIMENT_START_NATIVE_LANGUAGES_EQUALS = 1036,
  EXPERIMENT_START_NATIVE_LANGUAGES_NOT_EQUALS = 1037,

  /* Experiment: start - Learned languages */
  EXPERIMENT_START_LEARNED_LANGUAGES_EQUALS = 1038,
  EXPERIMENT_START_LEARNED_LANGUAGES_NOT_EQUALS = 1039,

  /* Variable */
  VARIABLE_CAN_NOT_SED_TO_SLIDE = 1040,
  VARIABLE_ARRAY_LENGTH_NOT_CORRECT = 1041,
  VARIABLE_NOT_ALL_COLUMNS_SET = 1042,
  VARIABLE_CAN_NOT_REMOVE_ALL_ATTRIBUTES = 1043,

  EXPERIMENT_ERROR_WHILE_CHECK_ACCESS = 1044,
  EXPERIMENT_CAN_NOT_UPDATE_PUBLISHED = 1045,

  BAD_REQUEST_IN_ENTITY = 1046,

  NO_DATA_FOR_CURRENT_SLIDE = 1047,
  NO_DATA_FOR_PREVIOUS_SLIDE = 1048,
  NO_DATA_FOR_START_TIME_PREVIOUS_SLIDE = 1049,
}